server/requirements.txt
asgiref==3.7.2
black==24.2.0
click==8.1.7
Django==5.0.2
django-cors-headers==4.3.1
django-filter==23.5
django-js-asset==2.2.0
django-mptt==0.16.0
django-unfold==0.20.5
djangorestframework==3.14.0
isort==5.13.2
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
pytz==2024.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
